add add user show controller

+++ src/main/java/hr/tpopovic/myshowlist/application/domain/model/Score.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 
 public sealed interface Score {
 
-    record Rated(Integer value) implements Score {
+    record Rated(Short value) implements Score {
 
         public Rated {
             requireNonNull(value, "Score value cannot be null");
